fix empty samplesheet test
-s : file exists AND is not empty
! -s : file does not exist OR is empty, instead of NOT (file exists AND is not empty)
-> [[ -e file.txt && ! -s file.txt ]] should only return TRUE if the file exists AND is empty, which is the purpose of this check!

+++ scripts/novaseq/checkfornewrun.bash
@@ -52,7 +52,7 @@ for RUN_DIR in ${IN_DIR}/*; do
         if [[ ! -f ${RUN_DIR}/demuxstarted.txt ]]; then
 
             # start with a clean slate: remove empty sample sheets before continuing
-            if [[ ! -s ${RUN_DIR}/SampleSheet.csv  ]]; then
+            if [[ -e ${RUN_DIR}/SampleSheet.csv && ! -s ${RUN_DIR}/SampleSheet.csv  ]]; then
                 rm ${RUN_DIR}/SampleSheet.csv
             fi
 
@@ -62,7 +62,7 @@ for RUN_DIR in ${IN_DIR}/*; do
             fi
 
             # exit if samplesheet is still empty after running demux sheet fetch
-            if [[ ! -s ${RUN_DIR}/SampleSheet.csv ]]; then
+            if [[ -e ${RUN_DIR}/SampleSheet.csv && ! -s ${RUN_DIR}/SampleSheet.csv ]]; then
                 echo "Sample sheet empty! Exiting!" 1>&2
                 continue
             fi
